Доработка 8 лабораторной, была доделана фотогалерея
